StartGame logic in GameBoard
Some pretty shaky assumptions going on here, but it's the best I can do
for now. I have no idea where cities are located, and I still need the
list of CityToCityRoutes. Waiting on @davishyer for that. @Jacobtk the
new methods in GameBoard are the most important things to check for
quality here. The file I load in should probably be glanced at too
(DestinationCards.txt). I manually combed through the actual game cards
to make the file.